Makes BubbleSort into it's own class.
